# backend_gotransact/docs/swagger.yaml
basePath: /api
definitions:
...
  responses.UserResponse:
    properties:
      data:
        additionalProperties: true
        type: object
      message:
        type: string
      status:
        type: integer
    type: object
  structutils.PaymentRequest:
    properties:
      amount:
        type: string
      cardNumber:
        type: string
      cvv:
        type: string
      description:
        type: string
      expiryDate:
        type: string
    required:
    - amount
    - cardNumber
    - cvv
    - expiryDate
    type: object
host: localhost:8000
info:
  contact:
    email: sangadevshatwar143@gmail.com
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for a transaction system.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Go Transaction API
  version: "1.0"
paths:
...
  /protected/postpayment:
    post:
      consumes:
      - application/json
      description: Posts a payment request with user authentication
      parameters:
      - description: Authorization header with bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Payment details
        in: body
        name: payment
        required: true
        schema:
          $ref: '#/definitions/structutils.PaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Payment request accepted
          schema:
            $ref: '#/definitions/responses.UserResponse'
        "400":
          description: Invalid input or validation error
          schema:
            $ref: '#/definitions/responses.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.UserResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.UserResponse'
      summary: Post a payment request
      tags:
      - Transaction
...
swagger: "2.0"